meta-mender-tegra: settings for firmware updates
On the tegra platform firmware rollback's are blocked by bootloader
edk2-nvidia. Trying to go back in version will end up in a non working
update that is hard to detect the root cause of. This commit will add
settings to control this behavior from Mender perspective.

To be able to detect the installed firmware version the UEFI variable
SystemFwVersions-781e084c-a330-417c-b678-38e696380cb9 is used. This
variable is not documented in edk2-nvidia but what is found by test
is that is reports version of firmware in slot A and version in slot B.
See discussion in [1].

Version of the capsule tegra-bl.cap is put in a file tegra-bl.cap.version
in build time and will be set to L4T_VERSION.

Default without changing configuration the present behavior is kept. By
setting TEGRA_MENDER_INSTALL_ONLY_IF_DIFFERENT in the configuration
bootloader will only be installed if version is newer than the one in the
inactive slot. If versions are same only the rootfs will be updated and switched.
By setting TEGRA_MENDER_INSTALL_ONLY_IF_DIFFERENT there will also be a check
if firmware version is older and the update will be rejected in Mender to
prevent later reject during UEFI capsule update.

Setting TEGRA_MENDER_ALLOW_FIRMWARE_VERSION_MISSMATCH to 1 will enable installation
of Mender update that contains an older version of the bootloader firmware.
Using this feature will make it possible to end up with non-official
combination of firmware and OS. For example when doing image updates
36.4.3(A)->36.4.4(B)->36.4.4(A)->36.4.3(B) you will end up running firmware
36.4.4 with the OS from 36.4.3 as this rollback protection will take effect in
the third upgrade.

[1] https://forums.developer.nvidia.com/t/edk2-nvidia-undocumented-variable-systemfwversions/344143

Signed-off-by: Peter Bergin <peter@berginkonsult.se>
Signed-off-by: Peter Bergin <peter.bergin@1x.tech>